requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.9.1
certifi==2025.11.12
distro==1.9.0
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
environ==1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.11.0
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
rest-framework-simplejwt==0.0.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
